refactor: personality 디렉토리 완전 제거
