--- agent/personality/__init__.py
@@ -1,11 +1,0 @@
-"""Personality 패키지 - 에이전트 성격 모듈."""
-
-from .elderly_companion import ElderlyCompanionAgent
-from ..prompts.greeting import CallDirection, GREETING_INBOUND, GREETING_OUTBOUND
-
-__all__ = [
-    "ElderlyCompanionAgent",
-    "CallDirection",
-    "GREETING_INBOUND",
-    "GREETING_OUTBOUND",
-]
