+++ agent/voice_agent.py
@@ -1,11 +1,11 @@
 """
-ElderlyCompanionAgent - 어르신 돌봄 AI 에이전트
+VoiceAgent - 어르신 돌봄 AI 에이전트
 
 메인 엔트리포인트: 모듈들을 조합하여 Agent 클래스 정의
 - AutoRAGMixin: 자동 과거 대화 검색
 - PipelineTimerMixin: 파이프라인 타이밍
 - GreetingManagerMixin: 인사말 관리
-- PersonaManagerMixin: 페르소나/프롬프트
+- Instructions Built via PromptBuilder (Persona Integrated)
 """
 
 import logging
@@ -13,19 +13,17 @@ from typing import Union
 
 from livekit.agents import Agent
 
-from ..pipeline_timer import PipelineTimerMixin
-from ..prompts.greeting import CallDirection, GreetingManagerMixin
-from ..tools.auto_rag import AutoRAGMixin
-from .persona_manager import PersonaManagerMixin
+from .pipeline_timer import PipelineTimerMixin
+from .prompts.greeting import CallDirection, GreetingManagerMixin
+from .tools.auto_rag import AutoRAGMixin
 
 logger = logging.getLogger(__name__)
 
 
-class ElderlyCompanionAgent(
+class VoiceAgent(
     AutoRAGMixin,
     PipelineTimerMixin,
     GreetingManagerMixin,
-    PersonaManagerMixin,
     Agent,
 ):
     """
@@ -38,7 +36,6 @@ class ElderlyCompanionAgent(
     - AutoRAGMixin: 자동 과거 대화 검색 및 컨텍스트 주입
     - PipelineTimerMixin: STT/LLM/TTS 타이밍 측정
     - GreetingManagerMixin: Redis Pub/Sub 인사말 관리
-    - PersonaManagerMixin: 시스템 프롬프트 생성
     """
 
     def __init__(
@@ -57,7 +54,7 @@ class ElderlyCompanionAgent(
             latitude: Ward's current location latitude (for weather tools)
             longitude: Ward's current location longitude (for weather tools)
         """
-        # Initialize prompt builder first (before calling _build_instructions)
+        self._ward_context = ward_context
         self._prompt_builder = None
 
         # Build instructions with persona (including location if available)
@@ -73,6 +70,26 @@ class ElderlyCompanionAgent(
         )
 
         logger.info(
-            f"ElderlyCompanionAgent initialized: "
+            f"VoiceAgent initialized: "
             f"direction={self.call_direction.value}, context_len={len(ward_context)}"
+        )
+
+    def _build_instructions(
+        self,
+        ward_context: str = "",
+        call_direction: Union[CallDirection, str] = CallDirection.INBOUND,
+        latitude: float | None = None,
+        longitude: float | None = None,
+    ) -> str:
+        """Build agent instructions using YAML-based prompt builder."""
+        # Lazy import to avoid circular dependency
+        if self._prompt_builder is None:
+            from .prompts.builder import PromptBuilder
+
+            self._prompt_builder = PromptBuilder(template_name="sodam")
+
+        return self._prompt_builder.build(
+            ward_context=ward_context,
+            latitude=latitude,
+            longitude=longitude,
         )
